requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0
langchain-qdrant
qdrant-client>=1.9.0
pypdf>=4.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
